# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Live Edge Engine API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GameStatus = (typeof GameStatus)[keyof typeof GameStatus];

export const GameStatus = {
  scheduled: "scheduled",
  live: "live",
  final: "final",
} as const;

export interface Team {
  id: string;
  name: string;
  abbreviation: string;
  logoUrl?: string;
  color?: string;
}

export interface Game {
  id: string;
  sport: string;
  status: GameStatus;
  startTime?: string;
  awayTeam: Team;
  homeTeam: Team;
  awayScore?: number;
  homeScore?: number;
  period?: string;
  clock?: string;
  isLive: boolean;
}

export type PlayerPropTrend =
  (typeof PlayerPropTrend)[keyof typeof PlayerPropTrend];

export const PlayerPropTrend = {
  up: "up",
  down: "down",
  flat: "flat",
} as const;

export type PlayerPropConfidence =
  (typeof PlayerPropConfidence)[keyof typeof PlayerPropConfidence];

export const PlayerPropConfidence = {
  High: "High",
  Medium: "Medium",
  Low: "Low",
} as const;

export type PlayerPropRecommendation =
  (typeof PlayerPropRecommendation)[keyof typeof PlayerPropRecommendation];

export const PlayerPropRecommendation = {
  Strong_Over: "Strong Over",
  Lean_Over: "Lean Over",
  Avoid: "Avoid",
  Lean_Under: "Lean Under",
  Strong_Under: "Strong Under",
} as const;

export type PlayerPropAction =
  (typeof PlayerPropAction)[keyof typeof PlayerPropAction];

export const PlayerPropAction = {
  Strong_Play: "Strong Play",
  Lean: "Lean",
  Avoid: "Avoid",
  Trap_Line: "Trap Line",
} as const;

export interface PlayerProp {
  id: number;
  sport: string;
  playerName: string;
  playerImage?: string;
  teamAbbr: string;
  teamLogo?: string;
  opponentAbbr: string;
  opponentLogo?: string;
  propType: string;
  line: number;
  avg5?: number;
  avg10?: number;
  hitRate5?: number;
  hitRate10?: number;
  lineGap?: number;
  consistency?: number;
  trend?: PlayerPropTrend;
  edgeScore: number;
  /** Model's estimated probability (0-100) that the recommended side hits. */
  winProbability: number;
  confidence?: PlayerPropConfidence;
  recommendation: PlayerPropRecommendation;
  action: PlayerPropAction;
  reasoning?: string;
  redFlags?: string[];
  riskWarning?: string;
  gameId?: string;
  /** Human-readable matchup label (e.g. "LAL @ DEN" or "NYY @ BOS") */
  gameLabel?: string;
  /** ISO timestamp of game start. */
  gameStartTime?: string;
  createdAt?: string;
}

export type LiveEdgeLiveRecommendation =
  (typeof LiveEdgeLiveRecommendation)[keyof typeof LiveEdgeLiveRecommendation];

export const LiveEdgeLiveRecommendation = {
  Strong_Live_Over: "Strong Live Over",
  Lean_Live_Over: "Lean Live Over",
  Avoid: "Avoid",
  Lean_Live_Under: "Lean Live Under",
  Strong_Live_Under: "Strong Live Under",
} as const;

export interface LiveEdge {
  id: number;
  propId: number;
  playerName: string;
  playerImage?: string;
  teamAbbr?: string;
  propType: string;
  line: number;
  currentStat: number;
  minutesPlayed?: number;
  percentComplete: number;
  projectedFinal: number;
  neededRemaining?: number;
  liveEdgePercent?: number;
  liveRecommendation: LiveEdgeLiveRecommendation;
  sport?: string;
  gameStatus?: string;
  period?: string;
  clock?: string;
  updatedAt?: string;
}

export type AlertSeverity = (typeof AlertSeverity)[keyof typeof AlertSeverity];

export const AlertSeverity = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface Alert {
  id: number;
  type: string;
  title: string;
  message: string;
  sport?: string;
  playerName?: string;
  edgeScore?: number;
  isRead: boolean;
  createdAt: string;
  severity: AlertSeverity;
}

export type ResultStatus = (typeof ResultStatus)[keyof typeof ResultStatus];

export const ResultStatus = {
  Pending: "Pending",
  Win: "Win",
  Loss: "Loss",
  Push: "Push",
  DNP: "DNP",
  Line_Removed: "Line Removed",
} as const;

export interface Result {
  id: number;
  date: string;
  sport: string;
  playerName: string;
  teamAbbr: string;
  opponentAbbr: string;
  propType: string;
  line: number;
  recommendation: string;
  edgeScore: number;
  status: ResultStatus;
  profitLoss?: number;
  closingLine?: number;
  createdAt?: string;
}

export type PatchResultBodyStatus =
  (typeof PatchResultBodyStatus)[keyof typeof PatchResultBodyStatus];

export const PatchResultBodyStatus = {
  Pending: "Pending",
  Win: "Win",
  Loss: "Loss",
  Push: "Push",
  DNP: "DNP",
  Line_Removed: "Line Removed",
} as const;

export interface PatchResultBody {
  status: PatchResultBodyStatus;
  profitLoss?: number;
  closingLine?: number;
}

export interface ResultsSummary {
  total: number;
  wins: number;
  losses: number;
  pushes: number;
  winRate: number;
  totalProfitLoss: number;
  roi: number;
}

export type ProviderStatusStatus =
  (typeof ProviderStatusStatus)[keyof typeof ProviderStatusStatus];

export const ProviderStatusStatus = {
  live: "live",
  mock: "mock",
  error: "error",
} as const;

export interface ProviderStatus {
  name: string;
  status: ProviderStatusStatus;
  lastChecked: string;
  description?: string;
  mockMode: boolean;
}

export interface DashboardSummary {
  totalGamesToday: number;
  liveGames: number;
  strongPlays: number;
  totalProps: number;
  avgEdgeScore: number;
  activeAlerts: number;
  recentWinRate: number;
  topEdgeProp?: PlayerProp;
}

export type GetScheduleTodayParams = {
  sport?: GetScheduleTodaySport;
};

export type GetScheduleTodaySport =
  (typeof GetScheduleTodaySport)[keyof typeof GetScheduleTodaySport];

export const GetScheduleTodaySport = {
  NBA: "NBA",
  MLB: "MLB",
  ALL: "ALL",
} as const;

export type GetScheduleToday200 = {
  games: Game[];
  lastUpdated: string;
};

export type GetLiveScores200 = {
  games: Game[];
  lastUpdated: string;
};

export type GetBestPropsParams = {
  sport?: GetBestPropsSport;
  minEdge?: number;
};

export type GetBestPropsSport =
  (typeof GetBestPropsSport)[keyof typeof GetBestPropsSport];

export const GetBestPropsSport = {
  NBA: "NBA",
  MLB: "MLB",
  ALL: "ALL",
} as const;

export type GetBestProps200 = {
  props: PlayerProp[];
  lastUpdated: string;
};

export type GetLiveEdge200 = {
  edges: LiveEdge[];
  lastUpdated: string;
};

export type GetAlertsParams = {
  unreadOnly?: boolean;
};

export type GetAlerts200 = {
  alerts: Alert[];
  unreadCount: number;
};

export type GetResultsParams = {
  sport?: string;
  status?: string;
};

export type GetResults200 = {
  results: Result[];
  summary: ResultsSummary;
};

export type RefreshData200 = {
  success: boolean;
  message: string;
};

export type GetApiStatus200 = {
  providers: ProviderStatus[];
  mockMode: boolean;
};