# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Live Edge Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /alerts:
    get:
      operationId: getAlerts
      tags: [alerts]
      summary: Get alerts
      parameters:
        - name: unreadOnly
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Alerts list
          content:
            application/json:
              schema:
                type: object
                properties:
                  alerts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Alert"
                  unreadCount:
                    type: number
                required: [alerts, unreadCount]
  /alerts/{id}/read:
    post:
      operationId: markAlertRead
      tags: [alerts]
      summary: Mark alert as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /results:
    get:
      operationId: getResults
      tags: [results]
      summary: Get results tracker
      parameters:
        - name: sport
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/Result"
                  summary:
                    $ref: "#/components/schemas/ResultsSummary"
                required: [results, summary]
  /results/{id}:
    patch:
      operationId: updateResult
      tags: [results]
      summary: Update a tracked result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchResultBody"
      responses:
        "200":
          description: Updated result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
...
components:
  schemas:
...
    Alert:
      type: object
      properties:
        id:
          type: number
        type:
          type: string
        title:
          type: string
        message:
          type: string
        sport:
          type: string
        playerName:
          type: string
        edgeScore:
          type: number
        isRead:
          type: boolean
        createdAt:
          type: string
        severity:
          type: string
          enum: [high, medium, low]
      required: [id, type, title, message, isRead, createdAt, severity]
    Result:
      type: object
      properties:
        id:
          type: number
        date:
          type: string
        sport:
          type: string
        playerName:
          type: string
        teamAbbr:
          type: string
        opponentAbbr:
          type: string
        propType:
          type: string
        line:
          type: number
        recommendation:
          type: string
        edgeScore:
          type: number
        status:
          type: string
          enum: [Pending, Win, Loss, Push, DNP, Line Removed]
        profitLoss:
          type: number
        closingLine:
          type: number
        createdAt:
          type: string
      required: [id, date, sport, playerName, teamAbbr, opponentAbbr, propType, line, recommendation, edgeScore, status]
    PatchResultBody:
      type: object
      properties:
        status:
          type: string
          enum: [Pending, Win, Loss, Push, DNP, Line Removed]
        profitLoss:
          type: number
        closingLine:
          type: number
      required: [status]
    ResultsSummary:
      type: object
      properties:
        total:
          type: number
        wins:
          type: number
        losses:
          type: number
        pushes:
          type: number
        winRate:
          type: number
        totalProfitLoss:
          type: number
        roi:
          type: number
      required: [total, wins, losses, pushes, winRate, totalProfitLoss, roi]
...